changed SyntaxCollectorVisitor to public helper class, added more test cases for integration test
