+++ src/Bicep.Core.IntegrationTests/ParamsParserTests.cs
@@ -1,13 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using Bicep.Core.UnitTests.Assertions;
 using Bicep.Core.Parsing;
 using Bicep.Core.Samples;
-using Bicep.Core.Syntax;
 using Bicep.Core.UnitTests.Utils;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -19,57 +17,24 @@ namespace Bicep.Core.IntegrationTests
     [TestClass]
     public class ParamsParserTests
     {
-        private class SyntaxCollectorVisitor : SyntaxVisitor
-        {
-            public record SyntaxItem(SyntaxBase Syntax, SyntaxBase? Parent, int Depth);
-
-            private readonly IList<SyntaxItem> syntaxList = new List<SyntaxItem>();
-            private SyntaxBase? parent = null;
-            private int depth = 0;
-
-            private SyntaxCollectorVisitor()
-            {
-            }
-
-            public static ImmutableArray<SyntaxItem> Build(ProgramSyntax syntax)
-            {
-                var visitor = new SyntaxCollectorVisitor();
-                visitor.Visit(syntax);
-
-                return visitor.syntaxList.ToImmutableArray();
-            }
-
-            protected override void VisitInternal(SyntaxBase syntax)
-            {
-                syntaxList.Add(new(Syntax: syntax, Parent: parent, Depth: depth));
-
-                var prevParent = parent;
-                parent = syntax;
-                depth++;
-                base.VisitInternal(syntax);
-                depth--;
-                parent = prevParent;
-            }
-        }
-
         [NotNull]
         public TestContext? TestContext { get; set; }
 
         [DataTestMethod]
-        [DynamicData(nameof(GetData), DynamicDataSourceType.Method, DynamicDataDisplayNameDeclaringType = typeof(DataSet), DynamicDataDisplayName = nameof(DataSet.GetDisplayName))]
+        [DynamicData(nameof(GetParamData), DynamicDataSourceType.Method, DynamicDataDisplayNameDeclaringType = typeof(DataSet), DynamicDataDisplayName = nameof(DataSet.GetDisplayName))]
         [TestCategory(BaselineHelper.BaselineTestCategory)]
         public void Params_Parser_should_produce_expected_syntax(DataSet dataSet)
         {
             if (dataSet.BicepParam is null)
             {
-                return;
+                throw new InvalidOperationException();
             }
 
             var program = ParamsParserHelper.ParamsParse(dataSet.BicepParam);
-            var syntaxList = SyntaxCollectorVisitor.Build(program);
+            var syntaxList = SyntaxCollectorVisitorHelper.SyntaxCollectorVisitor.Build(program);
             var syntaxByParent = syntaxList.ToLookup(x => x.Parent);
 
-            string getLoggingString(SyntaxCollectorVisitor.SyntaxItem data)
+            string getLoggingString(SyntaxCollectorVisitorHelper.SyntaxCollectorVisitor.SyntaxItem data)
             {
                 // Build a visual graph with lines to help understand the syntax hierarchy
                 var graphPrefix = "";
@@ -91,7 +56,7 @@ namespace Bicep.Core.IntegrationTests
                 };
             }
 
-            TextSpan getSpan(SyntaxCollectorVisitor.SyntaxItem data) => data.Syntax.Span;
+            TextSpan getSpan(SyntaxCollectorVisitorHelper.SyntaxCollectorVisitor.SyntaxItem data) => data.Syntax.Span;
 
             var sourceTextWithDiags = DataSet.AddDiagsToParamSourceText(dataSet, syntaxList, getSpan, getLoggingString);
             var resultsFile = FileHelper.SaveResultFile(this.TestContext, Path.Combine(dataSet.Name, DataSet.TestFileMainParamSyntax), sourceTextWithDiags);
@@ -103,7 +68,7 @@ namespace Bicep.Core.IntegrationTests
                 actualLocation: resultsFile);
         }
 
-        private static IEnumerable<object[]> GetData()
+        private static IEnumerable<object[]> GetParamData()
         {
             return DataSets.ParamDataSets.ToDynamicTestData();
         }
